See merge request ISPWA/intershop-pwa!641

+++ src/app/extensions/quoting/shared/product/components/product-add-to-quote-dialog/product-add-to-quote-dialog.component.ts
@@ -34,6 +34,7 @@ export class ProductAddToQuoteDialogComponent implements OnChanges {
   @Input() quote: QuoteRequest;
   @Input() quoteLoading = false;
 
+  @Output() deleteQuoteRequest = new EventEmitter<string>();
   @Output() updateQuoteRequest = new EventEmitter<{ displayName: string; description?: string }>();
   @Output() submitQuoteRequest = new EventEmitter<void>();
   @Output() updateItem = new EventEmitter<LineItemQuantity>();
@@ -65,17 +66,29 @@ export class ProductAddToQuoteDialogComponent implements OnChanges {
 
   /**
    * Throws updateItem event when onUpdateItem event trigggerd.
+   * Throws deleteQuoteRequest event when last items quantity is changed to '0'.
    * @param item Item id and quantity pair that should be changed
    */
   onUpdateItem(item: LineItemQuantity) {
-    this.updateItem.emit(item);
+    if (this.quote.items.length === 1 && item.quantity === 0) {
+      this.deleteQuoteRequest.emit(this.quote.id);
+      this.hide();
+    } else {
+      this.updateItem.emit(item);
+    }
   }
 
   /**
    * Throws deleteItem event when delete button was clicked.
+   * Throws deleteQuoteRequest event when last item will be deleted.
    */
-  onDeleteItem(itemId) {
-    this.deleteItem.emit(itemId);
+  onDeleteItem(itemId: string) {
+    if (this.quote.items.length === 1) {
+      this.deleteQuoteRequest.emit(this.quote.id);
+      this.hide();
+    } else {
+      this.deleteItem.emit(itemId);
+    }
   }
 
   /**
